refactor(cli): split over-cap test shard to satisfy §19
handlers_code_fixes_tests.rs was 2879 lines; §19 hard limit is 2000.

Split into:
- handlers_code_fixes_tests.rs (1497 lines): imports, make_server, and
  tests covering jsdoc, missing-imports, unresolved calls, and import
  candidate ranking
- handlers_code_fixes_tests_part2.rs (1425 lines): jsdoc @import fixes,
  type-only import merges, pnpm, async, interface, and JSX annotation
  tests

Both shards share a duplicated make_server() helper. The part2 module is
declared in handlers_code_fixes.rs via #[path] alongside the existing
tests and nested_pkg_tests modules.
